Add tmux-open-nvim plugin (#88)
tmux-open-nvim is a plugin that finds an instance of neovim running in the current tmux
window and opens a file there. It integrates well with tmux-fingers or tmux-open for when
a file path is in a pane and you want to open that file in neovim in an adjacent pane.